region save test + type cleaning
