+++ pyqtgraph_scope_plots/signals_table.py
@@ -373,6 +373,7 @@ class ColorPickerSignalsTable(ContextMenuSignalsTable, HasSaveRestoreModel):
 
     def _restore_model(self, model: BaseTopModel) -> None:
         super()._restore_model(model)
+        assert isinstance(data_model, ColorPickerDataStateModel)
         data_name_colors = [
             (data_name, QColor(data_model.color[0], data_model.color[1], data_model.color[2]))
             for data_name, data_model in model.data.items()
